integrated the god library and it compiles, so far so good. To make any valuable use of it, further planning is required. The intension of reimplementing a type system is to make it dynamic and to replace the nescessarity for reflection by type assertion, which performs much better, but also reduces complexity.
To really achieve that, further thought has to be put into it. The
public interface should in no way shape or form under no circumstances
ever expose an empty interface, have no need at all for assertion, or
convertion of a value returned by, or passed to it.

optional intensions are: keeping the error type exclusive to communicate
errors and not misuse it for type-/, boundary-/, or membership of a set
checking.  Use if/ok only publicly when aboulutely nescessary and where
it makes sense, which would be loop iteration and maybe as a tradeof, to
superunify an interface for types that need significantly different
methods, which might as well turn out to be a questionable approach in
the first place.

 zum Commit vorgemerkte Änderungen:
	gelöscht:       agiledoc.test
	geändert:       token.go
	neue Datei:     token_test.go
	geändert:       types.go

 Änderungen, die nicht zum Commit vorgemerkt sind:
	geändert:       vendor/github.com/emirpasic (geänderter Inhalt, unversionierter Inhalt)